+++ CHANGELOG.md
@@ -25,6 +25,13 @@
 * Clients using `client-utils` now automatically discover the compute replica group leader.
 * Change consensus interface to support multiple contracts.
 * Use Merkle Patricia tree for state storage.
+* Dockerfile includes truffle for solidity development.
+* Add LRU cache storage backend that can wrap any existing storage backend to add an
+  in-memory cache.
+* RPCs can be configured to time out, treating compute nodes that
+  don't respond in time as failing. To enable timeouts, pass
+  `--rpc-timeout SECONDS` in clients that use client-utils and
+  `--forwarded-rpc-timeout SECONDS` in ekiden-compute.
 
 # 0.1.0-alpha.4